+++ pyproject.toml
@@ -249,6 +249,8 @@ module = [
     "neptune.internal.hardware.gpu.gpu_monitor",
     "neptune.internal.hardware.hardware_metric_reporting_job",
     "neptune.internal.id_formats",
+    "neptune.internal.notebooks.comm",
+    "neptune.internal.notebooks.notebooks",
     "neptune.internal.init.model",
     "neptune.internal.init.model_version",
     "neptune.internal.init.project",
