+++ src/neptune/core/operation_processors/async_operation_processor/constants.py
@@ -19,7 +19,7 @@ __all__ = ["STOP_QUEUE_STATUS_UPDATE_FREQ_SECONDS", "STOP_QUEUE_MAX_TIME_NO_CONN
 import os
 
 from neptune.envs import NEPTUNE_SYNC_AFTER_STOP_TIMEOUT
-from neptune.internal.init.parameters import DEFAULT_STOP_TIMEOUT
+from neptune.internal.parameters import DEFAULT_STOP_TIMEOUT
 
 STOP_QUEUE_STATUS_UPDATE_FREQ_SECONDS = 30.0
 STOP_QUEUE_MAX_TIME_NO_CONNECTION_SECONDS = float(os.getenv(NEPTUNE_SYNC_AFTER_STOP_TIMEOUT, DEFAULT_STOP_TIMEOUT))
