+++ src/neptune/objects/model.py
@@ -46,7 +46,7 @@ from neptune.internal.backends.nql import (
 from neptune.internal.container_type import ContainerType
 from neptune.internal.exceptions import NeptuneException
 from neptune.internal.id_formats import QualifiedName
-from neptune.internal.init.parameters import (
+from neptune.internal.parameters import (
     ASYNC_LAG_THRESHOLD,
     ASYNC_NO_PROGRESS_THRESHOLD,
     DEFAULT_FLUSH_PERIOD,
